Make it all work.
Application env key `:rivet` must have the app name in it again so that
the config can be built correctly. migrations.exs, index.exs, and
archive.exs are not hidden files but found within the app's priv/rivet
folder. index.exs and archive.exs are in a subfolder named after the
model they migrate.